+++ WizBot.Core/Services/Database/Models/CustomReaction.cs
@@ -7,11 +7,19 @@ namespace WizBot.Core.Services.Database.Models
 {
     public class CustomReaction : DbEntity
     {
+        #region Unused
+        
+        [Obsolete]
         [NotMapped]
         public Regex Regex { get; set; }
+        [Obsolete]
         public ulong UseCount { get; set; }
+        [Obsolete]
         public bool IsRegex { get; set; }
+        [Obsolete]
         public bool OwnerOnly { get; set; }
+        
+        #endregion
         
         public ulong? GuildId { get; set; }
         public string Response { get; set; }